Fix crash caused by screensaver or display sleep
When the display goes to sleep (or the screensaver starts) the
macantumbra application crashes as the CGDisplayCreateImageForRect()
function returns NULL.

The fix checks to see if the image returned from
CGDisplayCreateImageForRect is NULL, in this case it handles the error
by not proceeding with the application logic. It then calls the same
function after a delay of some seconds.

With this fix the antumbra device is supposed to stop glowing
while the display is off, but the results depend on CGSHWCaptureDesktop,
called by the CGDisplayCreateImageForRect, which sometimes produces an
error. In that case the antumbra device glows at full brightness.

Signed-off-by: Georges Boumis <developer.george.boumis@gmail.com>