Fix crash caused by screensaver or display sleep
When the display goes to sleep (or the screensaver starts) the
macantumbra application crashes as the CGDisplayCreateImageForRect()
function returns NULL.

The fix checks to see if the image returned from
CGDisplayCreateImageForRect is NULL, in this case it handles the error
by not proceeding with the application logic. It then calls the same
function after a delay of some seconds. When the display wakes up then
the fix restarts the application logic.

Signed-off-by: Georges Boumis <developer.george.boumis@gmail.com>